shared/requirements.txt
# 共享依赖
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
asyncpg==0.29.0           # PostgreSQL 异步驱动
pydantic==2.9.2
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
aiofiles==24.1.0
redis==5.2.0
httpx==0.27.2
python-dotenv==1.0.1
alembic==1.13.3
psycopg2-binary==2.9.9    # PostgreSQL 同步驱动 (Alembic)
